requirements.txt
# Project dependencies


rapidfuzz
python-levenshtein
fuzzywuzzy

bitsandbytes==0.44.1
transformers==4.46.1
huggingface-hub==0.26.2

torch==2.5.1

langchain
langchain-community
sentence-transformers
chromadb

google-api-python-client
dateparser
deep-translator
langdetect

python-telegram-bot
numpy
pandas
tensorflow
scikit-learn
joblib
requests

sentencepiece
accelerate>=0.26.0